refactor: move Error into prql_ast
