+++ crates/prql-ast/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod expr;
 mod span;
 pub mod stmt;
