+++ crates/prql-compiler/src/semantic/context.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, fmt::Debug};
 
 use super::*;
-use crate::error::Span;
 use crate::ir::pl::*;
+use crate::Span;
 
 /// Context of the pipeline.
 #[derive(Default, Serialize, Deserialize, Clone)]
