+++ crates/prql-compiler/src/semantic/lowering.rs
@@ -6,7 +6,6 @@ use anyhow::Result;
 use enum_as_inner::EnumAsInner;
 use itertools::Itertools;
 
-use crate::error::{Error, Reason, Span, WithErrorInfo};
 use crate::generic::{ColumnSort, WindowFrame};
 use crate::ir::generic::{InterpolateItem, Range, SwitchCase};
 use crate::ir::pl::{self, Ident, Lineage, LineageColumn, PlFold, QueryDef, TupleField};
@@ -16,6 +15,7 @@ use crate::semantic::module::Module;
 use crate::semantic::write_pl;
 use crate::utils::{toposort, IdGenerator};
 use crate::COMPILER_VERSION;
+use crate::{Error, Reason, Span, WithErrorInfo};
 
 use super::context::{self, Context, DeclKind};
 use super::NS_DEFAULT_DB;
